operator,srv6: Allow locatorLen mismatching with the structure
Allow having locatorLen mismatches to the SID structure for certain
conditions. See code comments for concrete examples. The use case in
mind is allowing users to create a locator pool like below:

```
spec:
  prefix: "fd00:0:0:0::/48"
  locatorLenBits: 64
  structure:
    locatorBlockLenBits: 32
    locatorNodeLenBits: 16
    functionLenBits: 32
    argumentLenBits: 0
```

This allows users to allocate /64 for each node in the cluster while
advertising SRv6 VPN route with BGP with Cisco IOS-XR compatible
structure F321632 (a.k.a. Wide-LIB).

The actual allocation logic is still unchanged. It will be covered in
the subsequent commits.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>