sidmanager: Preserve locator prefix in SIDInfo
Now the locator prefix may not match with the SID structure. Preserve it
on SIDInfo by propagating locator on allocation. This is useful for BGP
to calculate SID transposition offset and length. For example, with /64
locator and structure [32, 16, 32, 0], we want to transpose lower 16bits
of the function. Transposition offset in this case is 64 and
transposition length is (32 + 16 + 32) - 64 = 16.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>