ci: enable cilium-dnsproxy image builds
Add the cilium-dnsproxy image build to the OSS image build workflows for
now instead of having to duplicate them all and keeping them in sync.
The matrix configuration isn't expected to change often, so conflicts
should rarely occur.

As a future improvement we might consider defining the matrix for all of
these separately and then load it like we e.g. do in
conformance-ginkgo.yaml. But that's for a future change.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>