egressgwha: Remove unneeded labelled continue
removeUnusedEgressRules loops over egress policy BPF map looking for
entries to be cleaned up.  If the current entry is still valid, the loop
goes ahead to inspect the next one.

A previous refactoring removed the need for multiple nested loops. Since
now there is just a single one, a `continue` is enough and is more
readable than a labelled continue.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>