gha: run enterprise workflows from stable branches on schedule
582f77967d8b ("Make ariane configuration mindful") moved the definition
of the Ariane triggers for enterprise workflows to a separate file for
mindfulness purposes. However, that caused a regression in stable
branches testing, as enterprise workflows are no longer run on schedule
through the ariane-scheduled workflow. Let's adapt that workflow to
account for both OSS and enterprise workflows.

The modification to the ariane-scheduled workflow is not mindful.
However, I personally think that it is acceptable because:

* The workflow itself is extremely stable and it is only updated when
  a new branch is created, or an old one becomes EOL.
* The workflow already includes non-mindful changes to account for
  different branch names.
* The creation of a separate workflow to trigger enterprise workflows
  only would be potentially subject to skews with respect to the already
  existing one, which can be instead easily caught in case of conflicts.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>