egressgwha: warn when topology.kubernetes.io/zone is missing
Currently, the operator emits an error if a node doesn't have
topology.kubernetes.io/zone when AZ affinity is enabled.
This commit changes the log level to warn because EGW can continue
working as expected in this situation although the nodes without
the label will be ignored. It's enough if we can inform the users
about it by the warning.

Also it causes a flake on e2e test(egress-gateway-ha-az-affinity)
After running egress-gateway-ha-az-affinity test, IEGP and
topology.kubernetes.io/zone label will be removed from all nodes.
If the operator triggers reconciliation when the label deletion has been
reflected while IEGP config hasn't, it emits the error and
check-log-errors/no-errors-in-logs fails.

Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>