(cee) bpf: complexity-tests: test combination of OSS and HA EgressGW
While the ideal config for CEE users is to only use the HA mode of EGW, we
also want to support a model where both OSS and HA are enabled
(allowing configuration via either CEGP or IEGP). This allows for a smooth
transition from OSS to CEE.

Therefore make sure that the datapath compiles with both code paths enabled
at the same time.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>