gha: workaround nodeport issue in mixed routing mode test
Cilium is currently affected by a bug breaking Pod to NodePort traffic
through an intermediate node in tunnel mode, when host-firewall is
enabled, KPR is enabled but SocketLB does not take effect. This matches
the "5" and "6" matrix entries. In these cases, let's ensure that the
clustermesh-apiserver gets scheduled on the same node that we target
through the NodePort as a temporary workaround until it gets fixed.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>