enterprise/cilium-cli: rename dnsproxy feature
Rename the feature from cilium-dnsproxy-deployed to cilium-dnsproxy-ha
because it not only coverd the cilium-dnsproxy DaemonSet being deployed
but also the HA proxy being enabled in the ConfigMap. This also matches
the naming used in the enterprise docs.

Also move a check out of extractExternalDNSProxyFeature so it's more
obvious that it depends on extractFromConfigMap being called before.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>