srv6: Workaround the race condition of informer vs tracker per resources
ad6ca03 introduces the fix for the race condition reported by
cilium/cilium#31010. However, it was not enough. It only waits for a
random single resource and closes the watching channel. Instead, we must
do that for all relevant resources used in the tests. Otherwise, the
same race condition happens for other resources.

Fixes: #3244
Fixes: #2742

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>