ci: lint FQDN proxy Helm chart
This was missed during the migration of the isovalent/fqdn-proxy where
the Helm chart validation was performed in its own workflow. Let's add
it to the existing enterprise-kind workflow on isovalent/cilium where
the FQDN proxy helm chart is used in a successive step.

Also make sure kubeconform is updated by renovate while at it.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>