ci: Add docker build output for FIPs-compliant images
This commit introduces docker image build workflows that produce an MVP
for images built with FIPs-compliant crypto libraries. This effort is a
result of an ask from Confluent. See isovalent/roadmap#1335. There are
two steps which must be taken in order to enable FIPs-compliant crypto
libraries for a binary:

1. GOEXPERIMENT=boringcrypto must be set as an environment variable.
2. An import to "crypto/tls/fipsonly" must take place.

Currently, our image build process is not conducive to implement a
mindful approach for the aforementioned requirements, therefore an
ad-hoc approach has been taken to create an MVP that can be passed to
the customer. Changes are in-progress to image build process to enable a
more mindful approach, however due to time constraints an intermediate
solution must be taken.

An additional requirement involves using Syft to generate SBOMs for
container images. This was implemented in OSS via https://github.com/cilium/cilium/pull/32307
and still undergoing backport, hence another reason why creating
separate workflows for this ticket was needed.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>