ci: add FQDN HA proxy upgrade compatibility workflow
This workflow runs compatibility tests for FQDN HA proxy between latest
stable cilium and cilium-dnsproxy versions on the respective branch and
the versions built from the PR. This replaces the `compat-test.sh`
compatibility test script that so far needed to be run manually upon
release. Parts of the script are extracted into separate helper scripts
invoked by the CI workflow.

The workflow performs the following steps:

  1. Install latest stable cilium and cilium-dnsproxy (patch or minor,
     depending on matrix mode) and run FQDN HA specific tests.

  2. Upgrade cilium to PR/HEAD and run FQDN HA specific tests.

  3. Upgrade cilium-dnsproxy to PR/HEAD and run FQDN HA specific tests.

  4. Downgrade cilium to stable and run FQDN HA specific tests.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>