k8s: iegp: add support for EGW-IPAM
Add support for EGW-IPAM with the new EgressCIDRs field in the IEGP
spec. The field is used to request egress IP allocations from a list of
CIDR blocks.  The IPs allocated are reported in the egress group Status,
specifically in a map that associates each allocated egress IP to its
target gateway node, identified by its node IP.  Finally, a new
Conditions field in the IEGP status reports information regarding the
allocations status.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>