egressgwha: Add address allocation from operator manager
Run a simple egress IP allocation algorithm for policy group that
specifies a list of valid egress CIDRs. The algorithm allocates the
available IPs sequentially, starting from the first IP in the pool, up
to the minimum between the number of active egress gateways and the size
of the pool.

In case of still-active gateways for which an IP allocation was already
carried out, the algorithm keeps those reservations, in order to
minimize the Egress IPs churn when gateways becomes unhealthy and then
healty again, or when the Cilium daemonset is rolled out.

The operator is also in charge of finding conflicts among all the egress
CIDRs requested in all the policies.  A CIDR marked as conflicting
cannot be used for IP allocations and thus will not be feeded to the
policy egress CIDRs pool.

The manager detects both "internal" conflicts, that is, conflicts among
CIDRs from the same policy and "external" conflicts, that is, conflicts
among CIDRs from two different policies.

In case of an external conflict, the most recent policy (as per the
CreationTimestamp value) is marked as conflicting. This is done in order
to avoid disruption to other previous allocations in case a faulty
policy is added.

The results of the allocation algorithm is reported as a series of
conditions in each policy that requested allocations.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>