hubble/observer: add a local node watcher
This commit introduce the observer LocalNodeWatcher, which cache the
local node information to be filled in Hubble flows.

Because the labels representation differ between the internal
node.LocalNode struct and Hubble flows (a map and a key=val slice,
respectively), we need to maintain a cache in order to avoid re-building
the labels slice for each flow. The LocalNodeWatcher aim to solve this
and can be hooked to the observer's OnGetFlows.

Signed-off-by: Alexandre Perrin <alex@isovalent.com>