bfd: Adapt MinRxInterval during Echo Function & fix Poll Sequence handling
This change makes sure that RequiredMinRxInterval is not less than one second
when the Echo Function is active, and goes back to the configured value when
it is not active.

Related to that, it fixes how Poll Sequence is handled when DesiredMinTxInterval
or RequiredMinRxInterval are changed - to ensure new values are not used
until the Poll Sequence is terminated under specific conditions
specified in RFC 5880.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>