egressgwha: Stop IEGP parsing in case of non-empty invalid EgressIP
EgressIP field of each IsovalentEgressGatewayPolicy egress group spec is
optional, but if specified, it is used to SNAT egress traffic for the
group.  Being an optional parameter, no error is logged in case the
conversion to netip.Addr fails, and the field is silently ignored.

To inform the user of the failure in setting the requested egress IP,
fail the IEGP parsing in case of an invalid non-empty egress IP.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>