egressgwha: Configure egress IP and route in the gateway node
Add support for reconciling the egress gateway network configuration as
per IEGP Egress Group Status section.

When the operator assigns an Egress IP to a gateway node, the agent:

- add the egress IP to the requested egress interface, or the default
  one if left empty
- add a rule to make the traffic with a source IP address equal to the
  egress IP use the routing table reserved for egress gateway IPAM
  (source-address policy based routing)
- add a route into the egress gateway IPAM routing table so that the
  datapath is able to send SNATed traffic to the selected egress
  interface with a BPF fib_lookup

After receiving an egw policy update, the agent updates the `egress-ips`
statedb table, and in turn a statedb reconciler updates the network
configuration as needed (IP, rule and route), based on the policy event
type (insert, update or delete). The reconciler guarantees that the
operation is retried in case of a failure. Also, if (all or part of) the
required configuration is already in place (e.g: the route or the rule
have alrady been created) the reconciler acknowledge it without making
any unnecessary change.

The usage of a separate routing table allows the egress-gateway IPAM
reconciler to avoid tinkering with the rest of the node network
configuration, including the routes not installed by Cilium. At the same
time, the reconciler collects all the stale egress IPs and routes
whenever an IEGP gets deleted.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>