cilium-dbg: minor clarifications to the clustermesh status output
Add the term remote to clarify that the number of clusters reported by
the cilium-dbg and kvstoremesh-dbg status commands do not include the
local one, regardless of whether it is included in the clustermesh
configuration or not. Similarly, let's replace the term failures with
reconnections, as failures has a negative connotation, but they are
actually expected to happen when the clustermesh-apiserver in the
given remote cluster is restarted.

Suggested-by: Bruno M. Custódio <bruno@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>