egressgwha: Replace string with netip.Addr in policy handling
Avoid using string as an intermediate type while handling IEGP events.
In Go, strings are immutable and does not have interning, thus they may
lead to increased pressure on heap allocations and garbage collection.
Instead, the netip.Addr type is internally implemented with a uint128
and a pointer to an interned valued, for better efficiency.

The values are kept as netip.Addr until the conversion to string is
needed, and that is carried out by specific helpers to improve
readability.

Moreover, keeping the original netip.Addr type ease the extensibility of
the current logic since all the netip.Addr specific methods are
available. This will be handy for Egress Gateway IPAM logic in the
subsequent commits.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>