bgpv2: Introduce OSS -> CEE reconcile parameter upgrader
In BGPv2 for CEE, we don't watch the IsovalentBGPXXX resources directly.
Instead, the operator creates the mirror OSS resource and the events for
the OSS resources triggers the reconciliation.

Also, in the agent, we implement enterprise-specific reconcilers as a
ConfigReconciler which takes OSS resource as an argument. Therefore, in
the enterprise-specific reconcilers, we need to resolve CEE resources
from OSS resources to get an actual configuration.

We introduce reconcilerParamsUpgrader object which implements a method
upgrade. It takes OSS ReconcileParams and "upgrades" it to
EnterpriseReconcileParams by fetching IsovalentBGPNodeConfig and find
the BGPInstance corresponds to the OSS instance we're currently
reconciling.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>
Signed-off-by: harsimran pabla <hpabla@isovalent.com>