remote-proxy: fix a nil pointer when `RemoveRestoredRules` is called
This commit fixed a panic when remote FQDN proxy `client` is not initialized.
The methods `GetRules`, `RemoveRestoredRules`, and `UpdateAllowed` were calling
client even if the connection with the remote proxy was not established.  We
now verify if the client is nil before calling the gRPC API.

Additionally, the `resetClient()` function has been removed and its code is
integrated directly in `Start()` function to simplify the logic. The
`connection` field is removed from `FQDNRemoteProxy` and used directly.

Signed-off-by: Donia Chaiehloudj <donia.cld@isovalent.com>