remote-proxy: Refactor RemoveRestoredRules to use a controller
When FQDN HA is not deployed, calls to RemoveRestoredRules fail. This function
will only be called again after an endpoints regeneration, resulting in a
window of time where rules may not accurately reflect the current state.

To address this, we are moving to a controller-based approach. This allows us
to enqueue the list of restored rules to be removed as soon as FQDN HA is
available.

There is a little optimization regarding the lock in `UpdateAllowed`. The map
holding the rules was locked until the controller finished its job.  Now we
deduplicate the list of rules to process.

In order to avoid returning errors in `cilium status`, we introduced a field
`hasFQDNProxyBeenReached` set to true only after connection was established and
restored endpoints have been removed properly.

Signed-off-by: Donia Chaiehloudj <donia.cld@isovalent.com>