enteprise/cilium-cli: Add egw-ipam test scenario
Add a scenario to test the egress-gateway IPAM feature.

The tests applies an IEGP where an Egress CIDR is specified. After
waiting for the operator to assign an Egress IP, and for the agent to
correclty configure the egress gateway BFP maps, a sequence of curl
requests validates that the traffic from the selected pods is correctly
masqueraded with the allocated Egress IP.

The test allows to specify the Egress CIDR to configure in the IEGP via
the option `--egress-cidr`.

In a kind cluster setup with a ipam subnet configured to
"172.18.0.0/16", the test can be run with:

cilium connectivity test --include-unsafe-tests --test "egress-gateway-ha-ipam" --egw-ipam-cidrs "172.18.0.8/30"

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>