ent/build-images: fetch artifacts with specific pattern
It seems that docker/build-push-action started to store artifacts on
GitHub. This sort of affected the digests of the image build process as
it timeout while trying to download these artifacts. To fix this issue
we will only download the artifacts with the pattern "*image-digest *"
which are the only artifacts relevant for the image digests.

Related: 112a16db2270 ("build-images: fetch artifacts with specific pattern")
Fixes: 51b52d81364c ("chore(deps): update docker/build-push-action action to v6")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>