nat/stats: add topk metrics implementation using natstats table.
This builds on top of the existing pkg/maps/nat/stats.Cell module
which provides a managed statedb.Table[NatMapStats] which contains
the top-n most used NAT tuples
(see: a8687767ef5d8daa5b94f0aeccab6a35c57edb2d for more details).

This works by watching for changes on entries on the nat-stats table,
which is used to manage a special top-k metric which contains the
current top-k nat-stat tuples.

When a nat-stat entry is removed, such as if it is no longer in the
top-n values or the tuples have been GC'd by nat GC then the manager
will remove any corresponding metrics to ensure that the cardinality
of the metric is always fixed with proportion to the nat-stats
`--nat-map-stats-entries` value (default: 32).

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>