contrib: fix cilium mesh fast kind configuration
with commit dc52072572 ("clustermesh: switch to surge upgrade
strategy.") helm install will block if there's any clustermesh apiserver
instance that is unavailable ("maxUnavailable: 0").

Unfortunately this doesn't work with the "fast" kind targets for cilium
mesh, as `make kind-cilium-mesh-install-fast` will result in some agent
and operator pods that are not ready (as they need the actual binaries
from the `make kind-image-enterprise-fast` target), which in turn will
make the clustermesh apiserver not ready, which in turn will block the
helm install/cilium install command.

As "maxUnavailable: 0" is irrelevant for testing purposes, restore the
old behaviour for the cilium mesh kind cluster.

Suggested-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>