cilium/API: add /v1enterprise/config endpoint for retrieving CEE options
If multinetwork is not enabled, the /v1enterprise/network/attachment
endpoint returns a 501. The enterprise CNI plugin makes a request to
this endpoint on CNI ADD and uses the response to determine whether the
feature is enabled, falling back on the OSS implementation if a 501 is
returned.

Although this is benign, it can cause false alerts in customer
environments, signaling that some server side error has occurred.

Rather than relying on this endpoint to determine the user's
configuration, let's add a /v1enterprise/config endpoint, similar to the
existing OSS /config endpoint, that can be used to determine the
configuration of enterprise-only features.

Currently the only feature exposed by this endpoint is multi-network,
but it can be expanded in the future as needed.

Fixes: #1086

Signed-off-by: Tim Horner <timothy.horner@isovalent.com>