egressgwha: Add pool data structure for IPAM
Add a pool data structure to support egress gateway IPAM. The pool is a
simple per-policy data structure to hold a list of CIDRs from which
egress IPs can be allocated from. It relies on the bitmap allocator from
pkg/ipam/service/allocator package and uses a contiguos allocation
strategy, that is, the first available address from the first non-empty
range is selected for allocation.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>