ci: Expand FIPs CI workflow to build images on pushes
This PR expands the FIPs CI workflow to build images on pushes to
main-ce and feature branches. The pattern used to implement this change
was based on the `Image CI Build`, with modifications made to account
for the FIPs workflow not having an image build cache.

Current effort is underway to modularize image building workflows to
prevent the need for the copying-and-pasting this kind of commit
performs. Until this effort is completed however, copying-and-pasting is
the best method we have.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>