srv6: lower log severity level for not found ep identity
Lower log severity level from error to warning for case when
identity labels can not be found when lookuing up in
the identity cache. This can happen e.g. for endpoints beeing
deleted, if the endpoint store has not been updated yet.
There is no functionality change as part of this change.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>