egressgwha: Use groupStatus type for internal representation
Use groupStatus type as internal representation for egress group status
and delay the creation of the v1.IsovalentEgressGatewayPolicyGroupStatus
instance just before the propagation of the update to the k8s
api-server.

This commit makes no functional changes, but it will be useful later
when the status will be furtherly updated with the result of the IPAM
Egress IPs allocation algorithm.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>