enterprise/fqdn: queue FQDN updates per-connection, replay on connect.
When we newly connect to an FQDN client, it starts out with no rules. We
need to send it the full rule-set. Thus, when it first starts, we must
"prime the pump" with the existing FQDN state.

To easily manage this lifecycle, this transitions the queues and update
controllers to be per-connection, rather than permanent. Even though
there will only ever be one connection, this is still a useful way of
tracking the data.

Additionally, add a way to dump the existing state of the local
FQDNProxy, so we can replay all previous events to the newly started
remote proxy. This state is already in the proxy, we just need to add an
enterprise_only function to dump it in a format we can use.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>