ci: Fix image tag logic in FIPs image release workflow
The FIPs image release workflow is failing due to a bug in the logic
used to determine the image tag. Thank you to @michi-covalent who
figured out that the if statement was incorrect, since `false`
evaluates as true in bash. To fix this bug, this commit copies in the
logic from the standard image release workflow. This workflow does not
have a workflow_dispatch trigger anyways, so there's no point in keeping
it in.

Example failed run: https://github.com/isovalent/cilium/actions/runs/9866724308

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>