bgpv2: Add BFD state reconciler for BGPv2
Adds a new enterprise BGPv2 reconciler that hard-resets BGP peering
upon BFD peering transition to the Down state, if BFD is configured
for the peer in the virtual router instance.

Also makes use of the paramUpgrader interface instead of the
direct use of the reconcileParamsUpgrader in the reconcilerv2 cell,
so that it can be mocked more easily.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>