Document merging staged Cilium enterprise documentation branch
More context is available in this slack thread
https://isovalent.slack.com/archives/C01717C3VL5/p1721665748998349

Basically, to prepare for upcoming production releases, we're
experimenting with creating staging branches for upcoming minor releases
in the cilium-enterprise-docs repository (eg: `cilium/v1.16-ce`) so we
can have a place to "stage" changes to the documentation for a "yet to
be released" product version. As an example, in Cilium 1.16-ce we're
adding Hubble RBAC support to Hubble Relay. Since this feature is only
available in v1.16-ce, we cannot merge it into the master branch of the
documentation repo until after v1.16-ce is released. That's where the
new `cilium/v1.16-ce` branch comes into play. It's a "staging" branch to
merge documentation for "to be released" features for the upcoming minor
version of a given product.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>