enteprise/cilium-cli: Add retry opts for egw IPAM tests
Egress Gateway IPAM reconciliation configures the selected egress
gateway interface with an egress IP taken from the egress CIDR specified
in the policy. In case of two quick policy updates in a row, an
interface might be assigned two different egress IPs, despite the
operator strive to keep the same allocations whenever possible. In that
case, it is possible that the external services ARP caches won't be
updated immediately, leading to a temporary inability for the traffic to
flow back to the egress gateway.  Despite being just a temporary effect,
this can make the egw ipam e2e test flaky.

To solve this, two parameters are added to control the number of retries
and the delay between them for the curl to the external echo service in
the egw ipam related tests.  Since the default values of those options
should be kept higher than the ones already present in the cilium-cli
OSS codebase, a separate set of enterprise-only flags are introduced,
meant to change the curl behavior in egw ipam tests only.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>