egressgwha: Fix IPAM for multiple IEGPs selecting the same gw
When the agent configures the routes for egress gateway IPAM, it does
not take into account the destination CIDRs, using the IPv4 default
route 0.0.0.0/0.

This leads to an issue when multiple IEGPs using IPAM selects the same
gateway. The agent reacts configuring both egress IPs and adding the
related routes, but it ends up overwriting the first route when
installing the second one. This happens because the route destination is
the same, that is, the IPv4 default route.

As an example, consider the two IEGPs:

```
apiVersion: isovalent.com/v1
kind: IsovalentEgressGatewayPolicy
metadata:
  name: iegp-sample-client
spec:
  azAffinity: disabled
  destinationCIDRs:
  - 172.18.0.5/32 # kind-worker3
  egressCIDRs:
  - 172.18.0.8/30
  egressGroups:
  - nodeSelector:
      matchLabels:
        kubernetes.io/hostname: kind-worker2
  selectors:
  - podSelector:
      matchLabels:
        io.kubernetes.pod.namespace: cilium-test
        kind: client
```

```
apiVersion: isovalent.com/v1
kind: IsovalentEgressGatewayPolicy
metadata:
  name: iegp-sample-client-2
spec:
  azAffinity: disabled
  destinationCIDRs:
  - 172.18.0.3/32 # kind-worker4
  egressCIDRs:
  - 172.18.0.20/30
  egressGroups:
  - nodeSelector:
      matchLabels:
        kubernetes.io/hostname: kind-worker2
  selectors:
  - podSelector:
      matchLabels:
        io.kubernetes.pod.namespace: cilium-test
        kind: client
```

Applying those two IEGPs leads to just one route installed in the
kind-worker2 gateway node:

```
$ ks exec -ti cilium-mhnc5 -- ip route show table 2050
default dev eth0 proto kernel src 172.18.0.20
```

In this case, the route related to the egress IP 172.18.0.20 overwrites
the one for 172.18.0.8.

The commit fixes this taking into account the Destination CIDRs
specified in the policy. That list is added into the egress-ips stateDB
table and the reconciler installs a route with <src = egressIP, dst =
destination CIDR> for each destination CIDR specified in the IEGP.

This allows to have all the routes in place to let the bpf fib_lookup in
the datapath select the correct egress interface as specified in the
policy.

Applying the same policies above now installs the correct routes in the
gateway:

```
$ ks exec -ti cilium-mhnc5 -- ip route show table 2050
172.18.0.5/32 dev eth0 proto kernel src 172.18.0.20
172.18.0.3/32 dev eth0 proto kernel src 172.18.0.8
```

Fixes: 2b22884cff ("egressgwha: Configure egress IP and route in the gateway node")

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>