datapath: wrap errors from WriteNodeConfig with context
Fixes: #34330

There are a few netlink calls in this function that can potentially return
EINTR. We're seeing this pop up in CI, but there's not enough context in
the error messages to attribute it to a specific syscall. Add this context
so we can properly fix it next time it occurs.

Signed-off-by: Timo Beckers <timo@isovalent.com>