enterprise: avoid hive dependency cycle in mesh
The loader recently started depending on the BPF maps itself, but this
introduced a hive dependency cycle in the enterprise cilium mesh code,
since the cilium mesh policy map depends on the loader for
initialisation ordering.

Fix this by registering a reference to the loader via an invoke instead
of directly depending on it.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>