cilium-cli: fix clustermesh endpointslice sync test
Currently, the clustermesh endpointslice synchronization test part of
the connectivity suite does not work as expected, because dig always
returns a 0 exit code regardless of whether an entry for the given
address is found or not. Apparently, there's no way to tune this
behavior by means of options.

Let's workaround this by using nslookup instead, which exits with
code 1 in case of NXDOMAIN. Differently from dig, nslookup takes into
account search domains by default, which is the desired behavior in
this context. Finally, let's fix the queried address to be in the
service-name.service-namespace form, as incorrect previously.

Fixes: b081bea18060 ("connectivity: add endpointslice clustermesh sync test")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>