egressgwha: only call deriveFromGroupConfig() for relevant egressGroup
When processing an IEGP, the EGW-HA manager on each agent walks the
policy's egressGroups and collects their active & healthy gatewayIPs. If
the local node is selected as gateway by an egressGroup, we call
deriveFromGroupConfig() to determine the network parameters (egressIP and
egressInterface). But since https://github.com/isovalent/cilium/pull/242
the check for "is local node selected by the egressGroup?" is applied
against the set of *all* selected gateways, regardless which egressGroup
contributes it.

For a policy with multiple egressGroups (eg0, eg1, ...) where eg0 selects
the local node, we therefore call deriveFromGroupConfig() *twice*. First
when processing eg0, then mistakenly a second time when processing eg1.
This results in an error ("local node selected by multiple egress
gateway groups from the same policy").

Fix this by matching the local node against only the nodes selected by
the current egressGroup.

Note that the error is currently harmless - what matters is the first call
to deriveFromGroupConfig(), and that's triggered by the egressGroup that
actually selects the local node.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>