egressgwha: Remove gateways without egress IP from active list
When using egress-gateway IPAM feature, the allocation of egress IPs to
active gateways can fail for different reasons. As an example, the IP
pool specified in the IEGP could not be wide enough to assign a
different egress IP to each active gateway. In that case, we can end up
with a gateway that is listed as active in the IEGP Status section
("activeGatewayIPs" field) despite not having an egress IP assigned
("egressIPByGatewayIP" field).

This could lead to traffic leaving the cluster without being properly
masqueraded whenever the gateway without an egress IP allocated is
selected.  To avoid this, the commit remove all gateways without a valid
egress IP allocated from both the list of active gateways and the map of
active gateways by affinity zones.

To have a deterministic output and thus make unit-testing simpler, the
slice of active gateways is sorted before being written in the IEGP
Status section.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>