egressgwha: Enable egress-ips reconciler pruning
Enable prune operation in egress-ips reconciler. The current Prune
implementation compares the existing rules and routes in the node
routing policy database with the desired state in the egress-ips stateDB
table, removing unneeded entries.

This operation is safe since the reconciler takes into account only the
rules and routes previously installed by Cilium to support the egress
gateway IPAM feature. This is done adding a filter for table with index
2050 (reserved for egress gateway IPAM routing table index) when
querying for routes and rules.

Conversely, the reconciler does not delete any address assigned to the
network interfaces, since we do not have a way to reliably differentiate
between addresses added for egress-gateway IPAM and others added
externally.

The consequence of this is that Cilium might leak egress IPs in case of
a restart.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>