egressgwha: Postpone pruning after k8s caches sync
Gate the pruning operation until all the k8s caches watched by egress
gateway manager are synced. This is done to avoid pruning too
aggressively at startup when the snapshot in the stateDB does not yet
reflect the latest status of IsovalentEgressGatewayPolicy, CiliumNode
and CiliumEndpoint.

Besides waiting for the k8s caches to be synced, we also wait for the
first regeneration of the gateway configurations, where the stateDB
egress-ips table is actually loaded with the egress IPs allocated by the
operator.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>