enterprise/cilium-cli: Wait for network config in egw-ipam tests
When executing the egress-gateway tests related to the IPAM feature,
wait for the expected network configuration to be reconciled in all the
gateway nodes.  Specifically, the test now verifies that the allocated
egress IP has been assigned to an interface and the related rule and
route have been added to the routing policy database.

This allows to avoid flakiness in case of a temporary error from the
egw-ipam stateDB reconciler when updating the gateway nodes
configuration.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>