ipsec,encap: remove ipsec encrypt hooks from encap.h
Traditionally, when IPsec and tunnelling is enabled the IPsec encrypt
hooks, which set the appropriate marks on an skb for encryption, were
invoked on pod egress when the datapath understands the `skb` is
destined for tunnel delivery.

To support a move for Encrypted Overlay, where we move these encryption
hooks to the egress device, the initial step is to remove these
encryption hooks from encap.h where our tunnel-redirect logic exists.

This removal of IPsec encrypt hooks allow the pod egress traffic
destined for tunnel delivery to simply redirect to the tunnel device
like ordinary pod-to-pod traffic.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>