ipsec,encap: make pod-egress encryption hooks conditional
Traditionally, when IPsec and tunnelling is enabled the IPsec encrypt
hooks, which set the appropriate marks on an skb for encryption, were
invoked on pod egress when the datapath understands the `skb` is
destined for tunnel delivery.

To support a move to performing IPsec encryption at the host egress device, the
initial step is to make these pod egress hooks conditional.

Now, the EncryptedOverlay flag is borrowed to determine whether the
legacy pod egress or the new (to be added in a subsequent commit) host
egress hook.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>