clustermesh: fix incorrect global service to phantom service transition
Currently, the service cache logic which handles the transition from
global to phantom service is affected by a bug, which possibly leads
to stale backends in the following scenario:

1. The remote cluster advertised a given service foo/bar as global;
2. Either the foo/bar service does not exist in the local cluster, or
   it is not marked as global;
3. The remote cluster changes service foo/bar from global to phantom;
4. The local cluster creates service foo/bar and marks it as global;

At this point, the local foo/bar service will incorrectly include the
backends from the remote cluster as well. Note that this issue only
occurs if the transition at step 3. is perceived as atomic from the
point of view of the local cluster (e.g., if the service is first
marked as phantom, and subsequently the global annotation is removed).

Let's fix this by always removing the external endpoints from the global
service cache if the remote service is of type phantom.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>