egressgwha: Send gratuitous ARP when configuring egress IP
When the egress gateway IPAM allocation algorithm is triggered by an
update of an IEGP, it is possible that an egress IP previously allocated
to a node is handed off to another node. In this case, there might be an
interval where traffic is disrupted because of the stale ARP cache
entries of the remote services.

To minimize this interval, the agent sends a gratuitous ARP reply each
time it adds a new egress IP to the selected egress interface.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>