ipsec: unconditionally create and despecify VXLAN-in-ESP XFRM policies
As a prerequisite for moving IPsec encryption hooks to the egress device,
Cilium should equip a node in tunnel mode with XFRM policies for both
ESP-in-VXLAN and VXLAN-in-ESP encryption and decryption, simultaneously.

This is largely to support the roll out of VXLAN-in-ESP to a cluster.

While the roll out occurs a node may see both VXLAN-in-ESP and
ESP-in-VXLAN traffic sent to it.

This would occur, for example, in a 3 node cluster where only a single
node has been updated to send VXLAN-in-ESP traffic.

Likewise, in upgrade scenarios there maybe v1.16 cluster nodes which
were not upgraded yet. The same scenario applies.

Additionally, we can stop specifying the VXLAN-in-ESP as encrypted
overlay. This was done more-or-less as a hack at the time for Multicast.
After some reflection there should be no reason to treat this traffic as
specify and identify it as such. Therefore, we stop utilizing specified
output marks and reqIDs.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>