ipsec: unconditionally create VXLAN-in-ESP related policies and states
Unconditionally create the XFRM policies and states which decrypt
VXLAN-in-ESP traffic, regardless of the state of the
EnableIPSecEncryptedOverlay config option.

This change allows a Cilium node to decrypt both VXLAN-in-ESP and
ESP-in-VXLAN traffic.
This will enable Cilium nodes to handle an upgrade scenario in the
future where VXLAN-in-ESP becomes the default.

While this note exists in the code's comment its important enough to
include in the commit:

We make an additional IN policy which checks for mark 0 and is
optionally enforced.

This handles two additional traffic scenarios introduced by VXLAN-in-ESP
traffic.
1. When ESP traffic is decrypted, VXLAN is inside. Both the outter and
	  inner headers therefore share the same source and destination addresses,
   the InternalIPs.
   When XFRM sees the decrypted VXLAN traffic the skb mark is set to zero
   by Cilium's datapath, as a way to signal the decryption processing is
   done. The decrypted VXLAN packet however is subjected to XFRM policy
   lookup again. Therefore, we need a policy which ALSO matches on the
   zero mark, otherwise a policy lookup for this packet will be done and
   nothing will match.
2. When in tunnel mode node-to-node traffic shares the same source and
	  destination addresses as both ESP and VXLAN traffic.
	  Therefore, this traffic is evaluated by our XFRM policies as well.
	  node-to-node traffic is passed to the stack by Cilium's datapath with
	  a mark set to zero. However, the traffic is NOT ESP. Therefore, the
	  same policy which fixes the above can be set to optional to allow
	  matching traffic with mark set to zero to not be enforced.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>